Added functionality to link to other page from tooltip.
Please note that this functionality only works on mobile devices. On
non-mobile devices, when the mouse hovers the popup, it will start to
disappear immediately, not allowing the user to click the popup.
This is caused by the pointer-events: all setting in css, which is
needed to enable the onclick event in the tooltip anyhow.